[aws-api] move S3Client into module
