+++ aws-api/src/main/scala/net/kemitix/s3thorp/aws/api/UploadProgressLogging.scala
@@ -1,6 +1,6 @@
-package net.kemitix.s3thorp.awssdk
+package net.kemitix.s3thorp.aws.api
 
-import net.kemitix.s3thorp.awssdk.UploadEvent.{ByteTransferEvent, RequestEvent, TransferEvent}
+import net.kemitix.s3thorp.aws.api.UploadEvent.{ByteTransferEvent, RequestEvent, TransferEvent}
 import net.kemitix.s3thorp.domain.LocalFile
 
 trait UploadProgressLogging {
