+++ src/main/scala/net/kemitix/s3thorp/ActionSubmitter.scala
@@ -2,8 +2,8 @@ package net.kemitix.s3thorp
 
 import cats.effect.IO
 import net.kemitix.s3thorp.Action.{DoNothing, ToCopy, ToDelete, ToUpload}
-import net.kemitix.s3thorp.S3Action.DoNothingS3Action
-import net.kemitix.s3thorp.awssdk.{S3Client, UploadProgressListener}
+import net.kemitix.s3thorp.aws.api.S3Action.DoNothingS3Action
+import net.kemitix.s3thorp.aws.api.{S3Action, S3Client, UploadProgressListener}
 import net.kemitix.s3thorp.domain.Config
 
 object ActionSubmitter {
