+++ src/main/scala/net/kemitix/s3thorp/SyncLogging.scala
@@ -1,6 +1,7 @@
 package net.kemitix.s3thorp
 
-import net.kemitix.s3thorp.S3Action.{CopyS3Action, DeleteS3Action, UploadS3Action}
+import net.kemitix.s3thorp.aws.api.S3Action.{CopyS3Action, DeleteS3Action, UploadS3Action}
+import net.kemitix.s3thorp.aws.api.S3Action
 import net.kemitix.s3thorp.domain.Config
 
 // Logging for the Sync class
