+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientDeleter.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp.awssdk
 
 import cats.effect.IO
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
-import net.kemitix.s3thorp.S3Action.DeleteS3Action
+import net.kemitix.s3thorp.aws.api.S3Action.DeleteS3Action
 import net.kemitix.s3thorp.domain.{Bucket, Config, RemoteKey}
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest
