+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientMultiPartUploader.scala
@@ -5,8 +5,8 @@ import cats.implicits._
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model._
 import net.kemitix.s3thorp.MD5HashGenerator.md5FilePart
-import net.kemitix.s3thorp.S3Action.{ErroredS3Action, UploadS3Action}
-import net.kemitix.s3thorp._
+import net.kemitix.s3thorp.aws.api.S3Action.{ErroredS3Action, UploadS3Action}
+import net.kemitix.s3thorp.aws.api.{S3Action, UploadProgressListener}
 import net.kemitix.s3thorp.domain.{Bucket, LocalFile, MD5Hash}
 
 import scala.collection.JavaConverters._
